code/requirements.txt
cos_python_sdk_v5==1.9.25
fake_useragent==1.1.3
langchain==0.0.271
openai==0.27.2
pandas==2.0.3
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
regex==2023.6.3
streamlit==1.25.0
streamlit_chat==0.0.2.2
